introducing item factory, removed cyclic ts refs
